feat(firebase emulator): connect app to firebase emulators, future development will be through local emulators

+++ src/app/components/listing-edit/listing-edit.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { Firestore, deleteDoc, updateDoc, doc, collection } from '@angular/fire/firestore';
 import { deleteObject, listAll, ref, Storage } from '@angular/fire/storage';
 import { FirestoreCollections, ImageFileVersion } from 'src/app/shared/globals';
+import { environment } from 'src/environments/environment';
 import { Listing } from '../listing-search/listing-search.data';
 
 @Injectable({ providedIn: 'root' })
@@ -14,17 +15,19 @@ export class ListingEditService {
 
     /* Completely remove the listing from DB */
     async deleteListing(listing: Listing, dbRefId: string) {
-        const allImages = (await listAll(ref(this.storage, listing.imageFolderPath!))).prefixes;
-        await Promise.all(allImages.map(async image => {
-            await Promise.all(
-                [
-                    deleteObject(ref(image, ImageFileVersion.compressed)),
-                    deleteObject(ref(image, ImageFileVersion.raw))
+        if (environment.production) {
+            const allImages = (await listAll(ref(this.storage, listing.imageFolderPath!))).prefixes;
+            await Promise.all(allImages.map(async image => {
+                await Promise.all(
+                    [
+                        deleteObject(ref(image, ImageFileVersion.compressed)),
+                        deleteObject(ref(image, ImageFileVersion.raw))
 
-                ]
-            );
-        }));
-        await deleteDoc(doc(this.firestore, dbRefId));
+                    ]
+                );
+            }));
+        }
+        await deleteDoc(doc(this.firestore, `${FirestoreCollections.listings}/${dbRefId}`));
     }
 
     /* Change archived property of the listing */
